fix: API를 통하지 않고 realtime 서버가 직접 서버 정보를 등록하도록 수정
