--- api/src/main/kotlin/com/backgu/amaker/api/realtime/dto/response/RealTimeServerHealthResponse.kt
@@ -1,7 +1,0 @@
-package com.backgu.amaker.api.realtime.dto.response
-
-data class RealTimeServerHealthResponse(
-    val status: String,
-) {
-    fun isValid(): Boolean = status == "UP"
-}
