+++ infra/src/main/kotlin/com/backgu/amaker/infra/redis/realtime/repository/RealTimeServerRepository.kt
@@ -3,4 +3,4 @@ package com.backgu.amaker.infra.redis.realtime.repository
 import com.backgu.amaker.infra.redis.realtime.hash.RealTimeServerHash
 import org.springframework.data.repository.CrudRepository
 
-interface RealTimeServerRepository : CrudRepository<RealTimeServerHash, Long>
+interface RealTimeServerRepository : CrudRepository<RealTimeServerHash, String>
