--- realtime/src/main/kotlin/com/backgu/amaker/realtime/orchestration/dto/ServerInformation.kt
@@ -1,8 +1,0 @@
-package com.backgu.amaker.realtime.orchestration.dto
-
-data class ServerInformation(
-    val id: Long,
-    val schema: String,
-    val address: String,
-    val port: Int,
-)
